research/neural_clustering/requirements.txt
absl-py
numpy
tensorflow-gpu==2.4.0
